suggestions/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.0
watchdog==6.0.0
google-genai==1.5.0